HW_2 - add enum and reorganize files
