+++ src/main/java/HW_2/Car.java
@@ -1,4 +1,4 @@
-package HW_2.task_2;
+package HW_2;
 
 public class Car implements Alarm, Vehicle{
     @Override
