+++ src/main/java/HW_2/Main.java
@@ -1,0 +1,11 @@
+package HW_2;
+
+import java.util.Arrays;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println(Arrays.toString(EnumCarMenu.values()));
+
+        
+    }
+}
