+++ src/main/java/HW_2/Mersedes.java
@@ -1,4 +1,4 @@
-package HW_2.task_2;
+package HW_2;
 
 public class Mersedes extends Car{
     String brand;
