--- src/main/java/HW_2/task_2/EnumMenu.java
@@ -1,4 +1,0 @@
-package HW_2.task_2;
-
-public class EnumMenu {
-}
